Decoupled headers that use external libraries (FFT and sndfile), so that client applications don't have to install those libraries unless necessary.
