+++ aku/AudioReader.cc
@@ -1,13 +1,16 @@
 #include <limits.h>
-#include "AudioReader.hh"
 #include <string>
 #include <iostream>
 #include <fcntl.h>
+#include <sndfile.h>
+
+#include "AudioReader.hh"
 
 namespace aku {
 
 AudioReader::AudioReader()
-  : m_sndfile(NULL),
+  : sf_info(new SF_INFO),
+    m_sndfile(NULL),
     m_little_endian(true),
     m_raw(false),
     m_eof_sample(INT_MAX),
@@ -16,11 +19,12 @@ AudioReader::AudioReader()
     m_end_sample(-INT_MAX),
     m_file_sample(0)
 {
-  sf_info.format = 0;
+  sf_info->format = 0;
 }
 
 AudioReader::AudioReader(int buffer_size)
-  : m_sndfile(NULL),
+  : sf_info(new SF_INFO),
+    m_sndfile(NULL),
     m_little_endian(true),
     m_eof_sample(INT_MAX),
     m_buffer_size(0),
@@ -28,7 +32,7 @@ AudioReader::AudioReader(int buffer_size)
     m_end_sample(-INT_MAX),
     m_file_sample(0)
 {
-  sf_info.format = 0;
+  sf_info->format = 0;
 }
 
 AudioReader::~AudioReader()
@@ -43,7 +47,7 @@ AudioReader::reset()
   m_start_sample = -INT_MAX;
   m_end_sample = -INT_MAX;
   m_file_sample = 0;
-  sf_info.format = 0;
+  sf_info->format = 0;
 }
 
 void // private
@@ -85,19 +89,19 @@ AudioReader::open(const char *filename, int sample_rate)
   reset();
 
   if (!m_raw)
-    m_sndfile = sf_open(filename, SFM_READ, &sf_info);
+    m_sndfile = sf_open(filename, SFM_READ, sf_info.get());
   
   if (m_sndfile == NULL)
   {
     // Try opening in RAW mode
-    sf_info.format = SF_FORMAT_RAW | SF_FORMAT_PCM_16;
+    sf_info->format = SF_FORMAT_RAW | SF_FORMAT_PCM_16;
     if (m_little_endian)
-      sf_info.format |= SF_ENDIAN_LITTLE;
+      sf_info->format |= SF_ENDIAN_LITTLE;
     else
-      sf_info.format |= SF_ENDIAN_BIG;
-    sf_info.samplerate = sample_rate;
-    sf_info.channels = 1;
-    m_sndfile = sf_open(filename, SFM_READ, &sf_info);
+      sf_info->format |= SF_ENDIAN_BIG;
+    sf_info->samplerate = sample_rate;
+    sf_info->channels = 1;
+    m_sndfile = sf_open(filename, SFM_READ, sf_info.get());
     if (m_sndfile == NULL)
       throw std::string("AudioReader::open(): could not open file:") +
         std::string(filename);
@@ -112,25 +116,25 @@ AudioReader::open(FILE *file, int sample_rate, bool shall_close_file, bool strea
   reset();
 
   if (!stream && !m_raw)
-    m_sndfile = sf_open_fd(fileno(file), SFM_READ, &sf_info, shall_close_file);
+    m_sndfile = sf_open_fd(fileno(file), SFM_READ, sf_info.get(), shall_close_file);
   
   if (m_sndfile == NULL)
   {
     if (stream)
-      sf_info.seekable = 0;
+      sf_info->seekable = 0;
     else {
-      sf_info.seekable = 1;
+      sf_info->seekable = 1;
       rewind(file); // FIXME: How about non-seekable streams?
     }
     // Try opening in RAW mode
-    sf_info.format = SF_FORMAT_RAW | SF_FORMAT_PCM_16;
+    sf_info->format = SF_FORMAT_RAW | SF_FORMAT_PCM_16;
     if (m_little_endian)
-      sf_info.format |= SF_ENDIAN_LITTLE;
+      sf_info->format |= SF_ENDIAN_LITTLE;
     else
-      sf_info.format |= SF_ENDIAN_BIG;
-    sf_info.samplerate = sample_rate;
-    sf_info.channels = 1;
-    m_sndfile = sf_open_fd(fileno(file), SFM_READ, &sf_info, shall_close_file);
+      sf_info->format |= SF_ENDIAN_BIG;
+    sf_info->samplerate = sample_rate;
+    sf_info->channels = 1;
+    m_sndfile = sf_open_fd(fileno(file), SFM_READ, sf_info.get(), shall_close_file);
     if (m_sndfile == NULL)
       throw std::string("AudioReader::open(): sf_open_fd() failed");
   }
@@ -141,14 +145,14 @@ AudioReader::open(FILE *file, int sample_rate, bool shall_close_file, bool strea
 void // private
 AudioReader::check_audio_parameters()
 {
-  if (sf_info.channels != 1)
+  if (sf_info->channels != 1)
     throw std::string("AudioReader: sorry, audio files with multiple "
 		      "channels not supported");
 
-  if ((sf_info.format & SF_FORMAT_SUBMASK) != SF_FORMAT_PCM_16)
+  if ((sf_info->format & SF_FORMAT_SUBMASK) != SF_FORMAT_PCM_16)
     fprintf(stderr, "WARNING: AudioReader::check_audio_parameters(): "
 	    "sample format not SF_FORMAT_PCM_16 but 0x%08x\n",
-	    sf_info.format);
+	    sf_info->format);
 }
 
 void
@@ -160,14 +164,14 @@ AudioReader::close()
   if (sf_close(m_sndfile) != 0)
     throw std::string("AudioReader::close(): sf_close() failed\n");
   m_sndfile = NULL;
-  sf_info.format = 0;
+  sf_info->format = 0;
 }
 
 void // private
 AudioReader::read_from_file(int start, int end)
 {
   if (start != m_file_sample) {
-    if (sf_info.seekable)
+    if (sf_info->seekable)
       seek(start);
     else
       throw std::string("Trying to seek a non-seekable FILE\n");
@@ -225,6 +229,18 @@ AudioReader::fetch(int start, int end)
     read_from_file(std::max(start, old_end_sample), end);
 }
 
+int
+AudioReader::sample_rate() const
+{
+  return sf_info->samplerate;
+}
+
+int
+AudioReader::num_samples() const
+{
+  return sf_info->frames;
+}
+
 void // private
 AudioReader::seek(int sample)
 {
@@ -234,7 +250,7 @@ AudioReader::seek(int sample)
     return;
   if (sample >= m_eof_sample)
     return;
-  if (sf_info.seekable == 0)
+  if (sf_info->seekable == 0)
     throw std::string("AudioReader::seek(): non-seekable file");
   sf_count_t ret = sf_seek(m_sndfile, sample, SEEK_SET);
   if (ret < 0)
