+++ aku/lda.cc
@@ -8,6 +8,7 @@
 #include "conf.hh"
 #include "HmmSet.hh"
 #include "FeatureGenerator.hh"
+#include "FeatureModules.hh"
 #include "Recipe.hh"
 #include "SpeakerConfig.hh"
 #include "util.hh"
